--- packages/builder/src/pages/builder/app/[application]/automation/_flagged/index.old.svelte
@@ -1,10 +1,0 @@
-<script>
-  import { redirect } from "@roxi/routify"
-  import { automationStore } from "@/stores/builder"
-
-  $: {
-    if ($automationStore.automations?.length) {
-      $redirect(`./${$automationStore.automations[0]._id}`)
-    }
-  }
-</script>
