--- packages/builder/src/pages/builder/app/[application]/automation/_flagged/layout.new.svelte
@@ -1,1 +1,0 @@
-<slot />
