+++ packages/builder/src/pages/builder/app/[application]/automation/_layout.svelte
@@ -1,12 +1,1 @@
-<script lang="ts">
-  import { featureFlags } from "@/stores/portal"
-  import NewLayout from "./_flagged/layout.new.svelte"
-  import OldLayout from "./_flagged/layout.old.svelte"
-
-  $: layout = $featureFlags.WORKSPACES ? NewLayout : OldLayout
-</script>
-
-<!-- routify:options index=3 -->
-<svelte:component this={layout}>
-  <slot />
-</svelte:component>
+<slot />
